requirements.txt
Pillow~=9.1.1
scikit-learn==0.24.2
torch==1.9.0
torchvision==0.10.0
numpy~=1.22.4
matplotlib~=3.5.2
seaborn==0.11.1
